Change num_ratings to an IntegerField

+++ Quotebook/app/models.py
@@ -11,7 +11,7 @@ class Quote(models.Model):
     attribution = models.CharField(max_length=64)
 
     rating = models.FloatField(default=0)
-    num_ratings = models.FloatField(default=0);
+    num_ratings = models.IntegerField(default=0);
 
     def __str__(self):
         return 'Quote ' + str(self.pk) + ' by ' + self.attribution
